+++ NEWS.md
@@ -1,3 +1,8 @@
+# survivalAnalysis 0.1.1
+
+* export identity_order, which is used by forest plots
+* work around upstream name collision (flatten_raw in purrr and rlang)
+
 # survivalAnalysis 0.1.0
 
 * initial release